ueventd: Fix bootdevice by-name/by-num link creation
Create bootdevice by-name/by-num links only if the block
device is the bootdevice.  This fixes the issue of
bootdevice path being created incorrectly if the first
reported device with by-name/by-num is not the bootdevice.

It also fixes devices in which the bootdevice property
doesn't exist, or the symlink is created by a later step in
the init sequence (such as init.rc itself)

Change-Id: I8e103e04194a425eeea6302c3be745557cbaf886