init: Don't crash the system due to an invalid mkdir argument
Prior to this change, if someone write

      mkdir /data/log/sar_cap system system

in one of their init.rc files, this causes init to get a exception
which causes init to exit with no error message to allow you to track
this down which then causes your kernel to panic.

Let's make this a little more user friendly by simply reporting
an error instead of silently crashing the whole damn system.

Change-Id: I82a85aa2e79b138bd4cccfeb1f967f37365b8e18