fs_mgr: zram: allow specifying stream count
This parameter allows specifying the number of parallel streams to be used in compression and decompression. This option was added in Linux 3.15, so it is not very common among devices currently on the market, but it has been backported to some kernels. This option must be set before zRAM is activated, as the kernel does not allow switching from single stream to multistream compression once zRAM is active.

Change-Id: I6cd5916d9d29e5fd945fcc133a5e569c8873ce10