init: builtins: ignore directory policy failure
Happens when using a new FBE container with existing data.
(Maybe also when re-using an old container with wiped data)

This is a try to prevent or even completely workaround the TWRP data
wipe after e. g. a data restore.

I think this happens because when TWRP restores data, it also sets
the corresponding policies. Android then tries to do the same but
somehow fails to do so (maybe because they are already set but android
thinks they are not) so it panics and reboots.

If you are lucky you should be able to get booted if you have this
commit instead of being forced into a data format.

Additionally, to prevent the system from wiping data via recovery,
add following flag to your TWRP (in case you maintain one):

https://github.com/xdevs23/android_device_oneplus_cheeseburger-twrp/blob/android-7.1/BoardConfig.mk#L151

Change-Id: I9202c04cd7e0a9bb201617d9bcb05116d7f6c16d
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>